chapter3_4_aianimals/search/model_development/similar_image_search/pyproject.toml
[tool.poetry]
name = "similar_image_search"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.10.6"
tensorflow = "^2.7.0"
numpy = "^1.22.0"
click = "^8.0.3"
mlflow = "^1.22.0"
Pillow = "^9.0.0"
pydantic = "^1.9.0"
scann = "^1.2.5"
tensorflow-recommenders = "^0.6.0"
tensorflow-hub = "^0.12.0"
tensorflow-datasets = "^4.5.2"
tensorboard = "^2.8.0"
hydra-core = "^1.1.1"
httpx = "^0.23.0"
nptyping = "^1.4.4"
scikit-learn = "^1.0.2"
psycopg2-binary = "^2.9.3"
dependency-injector = "^4.39.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
